Removed the AnotherSMBW, NewerHS and NewerSumSun Game Patches
I've been meaning to remove these patches for a while since they aren't being maintained and are rather redundant. Right now the editor is focused on just supporting NSMBW and NewerSMBW (as well as NewerDX if it's different in a significant enough way).

I was planning on restoring/fixing any issues with the other Game Patches and leaving them as a separate download on the releases page, but eh... If I haven't done it by now then I'm probably not going to ever.

Someone could always do it themselves tho